Create Scripture Package (#544)
* Create HTMLView package

* use same version of r-n everywhere, hopefully fixing tests.

* fix tests and such

* move scripture view to its own component

* Include scripture in storybook

* Add scripture to storybook

* Ensure package will publish storybook

* Remove packages that aren't needed in peer deps
