+++ packages/apollos-ui-htmlview/.babelrc
@@ -1,0 +1,3 @@
+{
+  "presets": ["module:metro-react-native-babel-preset"]
+}
