+++ packages/apollos-ui-htmlview/.npmignore
@@ -1,0 +1,2 @@
+*
+!lib/**
