+++ packages/apollos-ui-scripture/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@apollosproject/ui-passes",
+  "name": "@apollosproject/ui-scripture",
   "version": "0.8.6",
   "homepage": "https://github.com/ApollosProject/apollos-prototype#readme",
   "main": "src/index.js",
@@ -26,40 +26,26 @@
     "testMatch": [
       "<rootDir>/src/**/?(*.)(test|tests).{js,jsx,mjs}"
     ],
-    "setupFiles": [
-      "./jest.setup.js"
-    ],
     "modulePaths": [
       "<rootDir>/src/"
     ]
   },
   "peerDependencies": {
     "@apollos-project/ui-kit": "*",
+    "@apollos-project/ui-htmlview": "*",
     "react": "*",
-    "react-navigation": "*",
-    "react-native": "*",
-    "react-native-passkit-wallet": "*",
-    "rn-fetch-blob": "*"
-  },
-  "dependencies": {
-    "@apollosproject/ui-kit": "^0.8.6",
-    "graphql-tag": "^2.10.1",
-    "react-apollo": "2.5.5"
+    "react-native": "*"
   },
   "devDependencies": {
-    "@apollosproject/babel-preset-apollos": "^0.8.6",
-    "codecov": "3.1.0",
+    "@apollosproject/ui-storybook": "^0.8.6",
+    "metro-react-native-babel-preset": "^0.53.1",
+    "@storybook/react-native": "^4.0.7",
     "jest": "23.6.0",
     "jest-react-native": "18.0.0",
     "prop-types": "15.6.1",
-    "react": "16.6.3",
-    "react-dom": "16.6.3",
+    "react": "^16.6.3",
     "react-native": "0.58.5",
-    "react-native-linear-gradient": "2.4.0",
-    "react-native-passkit-wallet": "^0.1.4",
-    "react-native-svg": "6.5.2",
-    "react-test-renderer": "16.6.1",
-    "rn-fetch-blob": "^0.10.15"
+    "react-test-renderer": "16.6.1"
   },
   "publishConfig": {
     "access": "public"
