+++ packages/apollos-ui-scripture/src/Scripture/ScriptureItem.js
