+++ packages/apollos-ui-scripture/src/Scripture/ScriptureList.js
