+++ packages/apollos-ui-scripture/src/Scripture/ScriptureList.tests.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
 
-import Providers from 'apolloschurchapp/src/Providers';
+import { Providers } from '@apollosproject/ui-kit';
 
 import ScriptureList from './ScriptureList';
