+++ packages/apollos-ui-scripture/src/index.js
@@ -1,0 +1,5 @@
+import Scripture, { ScriptureList } from './Scripture';
+import ScriptureItem from './Scripture/ScriptureItem';
+
+export default Scripture;
+export { ScriptureList, ScriptureItem };
